+++ packages/studio/src/components/AppRatings/messages.tsx
@@ -1,0 +1,11 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  anonymous: 'Anonymous',
+  ratings: 'Ratings',
+  rateApp: 'Rate this app',
+  ratingSuccessful: 'You successfully rated this app!',
+  you: 'It’s you!',
+  loadError: 'There was a problem loading app ratings.',
+  retry: 'Retry',
+});