+++ packages/studio/src/components/ClientCredentials/messages.tsx
@@ -3,6 +3,8 @@ import { defineMessages } from 'react-intl';
 export default defineMessages({
   title: 'OAuth2 client credentials',
   explanation: 'OAuth2 client credentials allow applications to perform actions on your behalf.',
+  loadError: 'There was a problem loading your client credentials.',
+  retry: 'Retry',
   empty:
     'You currently haven’t registered any OAuth2 clients. If you don’t know what this is, you should probably leave this empty.',
   register: 'Register',