+++ packages/studio/src/components/JSONSchemaEditor/types.ts
@@ -3,10 +3,7 @@ import type { Definition } from 'typescript-json-schema';
 
 import type { NamedEvent } from '../../types';
 
-export interface CommonJSONSchemaEditorProps<
-  T = never,
-  S extends OpenAPIV3.BaseSchemaObject = OpenAPIV3.NonArraySchemaObject
-> {
+export interface CommonJSONSchemaEditorProps<T = never> {
   /**
    * Whether or not the editor is disabled.
    *
@@ -46,7 +43,7 @@ export interface CommonJSONSchemaEditorProps<
   /**
    * The schema used to render the form elements.
    */
-  schema: S | Definition;
+  schema: OpenAPIV3.SchemaObject | Definition;
 
   /**
    * The value used to populate the editor.